Move event store out off test
